backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
psycopg2-binary==2.9.10
requests==2.32.3
openai==1.58.1
google-generativeai==0.8.4
python-dotenv==1.0.1
pytest==8.3.4
pytest-django==4.9.0
pytest-asyncio==0.25.0